NvidiaDeepResearch/requirements.txt
openai
dotenv
PyYAML
requests
aiohttp
Pillow
ipython